fixing types for nautilus compile

+++ c/f2_primfunks.c
@@ -1264,7 +1264,7 @@ def_pcfunk1(identity, exp,
 #define PRIME_NUMBER__16_BIT 65521
 
 u64 raw__random(u64 max_value) {
-  int random_value = (uint)(rand() * PRIME_NUMBER__16_BIT) % max_value;
+  int random_value = (u64)(rand() * PRIME_NUMBER__16_BIT) % max_value;
   random_value = (random_value < 0) ? (-random_value) : random_value;
   return random_value;
 }
